MenuComponent Final Commit
