+++ src/App.tsx
@@ -5,7 +5,7 @@ import {
   Routes,
   Route,
 } from "react-router-dom";
-import { Button, ThemeContext, DropdownMenu } from "./components";
+import { Button, ThemeContext, MenuComponent } from "./components";
 
 function App() {
   const { changeColorPalette } = useContext(ThemeContext);
@@ -46,7 +46,7 @@ function Home() {
   return (
     <div className="grid h-screen place-items-center">
       <Button  primary  >text blah blah blah </Button>
-      <DropdownMenu data = {menuData}/>
+      <MenuComponent data = {menuData}/>
      </div>
   );
 }
