--- src/components/DropdownMenu/index.ts
@@ -1,1 +1,0 @@
-export { DropdownMenu } from "./DropdownMenu";
