+++ src/components/MenuComponent/index.ts
@@ -1,0 +1,1 @@
+export { MenuComponent } from "./MenuComponent";
